requirements.txt
Flask==3.0.2
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
Pillow==10.2.0